pyproject.toml
[project]
name = "atr-code"
version = "0.1.0"
description = "Repository containing code for an Automated Target Recognition (ATR) project for Dassault Aviation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.3.1",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.1.0",
    "scikit-image>=0.26.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
